* AbstractCommand.cc: casted timeout value to long long int. 	* ChunkedEncoding.cc: fixed memory leak. 	* PeerInteractionCommand.cc: casted timeout value to long long int. 	* SleepCommand.cc: casted timeout value to long long int. 	* Data.cc: fixed memory leak. 	* Data.cc: fixed toLLInt(). 	* BitfieldMan.cc: fixed memory leak. 	* TorrentMan.cc: initialized storeDir to ".". 	* TorrentMan.cc: fixed memory leak. 	* TorrentMan.cc: corrected file paths of splitted files. 	* PeerAbstractCommand.cc: casted timeout to long long int. 	* main.cc: added delete(req) and delete(te->diskWriter). 	* RequestSlot.cc: casted timeout value to long long int. 	* Request.cc: fixed memory leak. 	* PendingMessage.cc: make HAVE messages sent only when peer does not 	have the piece. 	* Peer.{h,cc}: added hasPiece(int index).

+++ src/PeerInteractionCommand.cc
@@ -361,7 +361,7 @@ void PeerInteractionCommand::keepAlive() {
   } else {
     struct timeval now;
     gettimeofday(&now, NULL);
-    if(Util::difftv(now, keepAliveCheckPoint) >= 120*1000000) {
+    if(Util::difftv(now, keepAliveCheckPoint) >= (long long int)120*1000000) {
       if(pendingMessages.empty()) {
 	peerConnection->sendKeepAlive();
       }
